Only retry on connection errors
This has the nice side effect of giving up immediately on any other sort
of errors.  It only does that because we have a single node set up via LVS.

Also fix some complaints about "the config 'host' is not found".  'host'
isn't a config.  It is a property.  I don't have any clue why, but it is.

A part of fixing this bug:
Bug: 57215

Change-Id: I42b592f1e3d773e9f2f88733a0496f8be6944dbf