Make MWElasticUtils::deleteByQuery harder to misuse
deleteByQuery was a generator function which is fairly unusual for
php code, at least within the mediawiki ecosystem. The code in
Translate extension was calling this method but never iterating the
generator, which means the function never actually started.

Change deleteByQuery to work as most mw devs would expect, and block
until it has completed. Rename the current implementation into a new
method, deleteByQueryWithStatus, which exposes the generator based
interface.

Bug: T219234
Change-Id: I155593ed9ca982b320f0c6421210e306775d00c4