Use an increasing delay in MWElasticUtils::deleteByQuery
5 seconds delay between delete and check is acceptable for use cases
that run this once or twice in a job, but for use cases that need to
run deleteByQuery hundreds of times all those 5s sleeps add up to
enough time to push over the job runner max timeout.

Start the delay at 50ms and increase by 50% up to the max delay of 5s.

Bug: T219234
Change-Id: Icccf869de0b448fe6b09e3c7e560bc7563b7adde