Don't fail connection after CURLE_PARTIAL_FILE response
This means the connection dropped before the full response was read,
likely some sort of network problem or elasticsearch shut down
mid-response. If the network truly failed or elasticsearch is gone the
retry should fail. Should be safe to retry directly, but for safety
and to prevent an infinite loop delegate retries to calling code.

Bug: T265056
Change-Id: I16e15fc8c10eb8a3202cb3874ee20f3846d37511