Don't disable the last connection after http error (part deux)
Previously we adjusted this connection handling to ensure that the
last available elasticsearch connection would not be disabled. We
had limited that to the CURLE_COULDNT_CONNECT error, but it turns out
we can arrive there in other ways as well. This time around we saw
CURLE_GOT_NOTHING causing this. It seems like there will be a variety
of curl codes that could return, adjust our handling to provide a
better guarantee that a connection still exists after leaving the
connection error handling.

Bug: T341058
Change-Id: Id2cf5d8b1e4d8e6086d1748cc00f98336e4c58a8